refactored gradle build scripts to precompiled plugins
