+++ buildSrc/build.gradle.kts
@@ -3,8 +3,11 @@ plugins {
 }
 
 repositories {
-    jcenter()
+    gradlePluginPortal()
+    mavenCentral()
 }
 
 dependencies {
+    implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:1.4.10")
+    implementation("com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4")
 }
