+++ buildSrc/src/main/kotlin/kebbuild.keb-published-kotlin-module.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("kebbuild.project-info")
+    id("kebbuild.repositories")
+    id("kebbuild.kotlin-module")
+    id("kebbuild.publication")
+    id("kebbuild.testing")
+}
