+++ buildSrc/src/main/kotlin/kebbuild.project-info.gradle.kts
@@ -1,0 +1,3 @@
+group = "org.kebish"
+version = "1.1-SNAPSHOT"
+description = "Library for browser tests implementing Page Object pattern and using Selenium"
