+++ buildSrc/src/main/kotlin/kebbuild.repositories.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    `java-base`
+}
+
+repositories {
+    mavenCentral()
+    jcenter()
+}
