+++ gradle.properties
@@ -1,1 +1,5 @@
 kotlin.code.style=official
+org.gradle.parallel=true
+org.gradle.caching=true
+org.gradle.jvmargs=-Dfile.encoding=UTF-8
+org.gradle.vfs.watch=true
